fix subclass, fix clan, updates skills, fix DM.

+++ trunk/L2JTeon/java/net/sf/l2j/gameserver/model/actor/stat/DoorStat.java
@@ -41,7 +41,7 @@ public class DoorStat extends CharStat
 	}
 
 	@Override
-	public final /*byte*/int getLevel()
+	public final byte getLevel()
 	{
 		return 1;
 	}
